dashborad Card View

+++ src/app/modules/posts/posts.component.ts
@@ -1,10 +1,10 @@
+import { DialogConfirmServicesService } from './../../services/dialog-confirm-services.service';
 import { DialogConfirmComponent } from './../dialog-confirm/dialog-confirm.component';
 import { FormEnseignantComponent } from './../form-enseignant/form-enseignant.component';
 import { EnsService } from './../../services/ens.service';
 import { NgModule } from '@angular/core';
 import { Component, OnInit } from '@angular/core';
 import { MatDialogModule, MatDialogConfig, MatDialog } from '@angular/material/dialog';
-import { Popup } from 'ng2-opd-popup';
 
 @Component({
   selector: 'app-posts',
@@ -15,15 +15,13 @@ export class PostsComponent implements OnInit {
   searchText:string;
   number:number=1;
   columnlist:string[]=['Nom','Prenom','Email','Cin','DateNais','Adresse','Telephnoe','DateEmb','Cnss','Salaire'];
-  constructor(public service:EnsService,public dialog:MatDialog,private popup:Popup) { }
-  popoverTitle:string="Supprimer Cet enseignant "
-  popoverMessage:string="Vous êtes sûr de vouloir supprimer cet enseignant ?"
-  confirmClicked=false;
-  cancelClicked=false;
+  constructor(public service:EnsService,public dialog:MatDialog,public dialogService:DialogConfirmServicesService) { }
   ngOnInit(): void {
     this.service.GetAll();
       }
       onOpen(){
+        console.log(this.service.form.value)
+        this.service.form.reset();
         const dialogConfig=new MatDialogConfig();
         dialogConfig.disableClose=true;
         dialogConfig.autoFocus=true;
@@ -33,30 +31,29 @@ export class PostsComponent implements OnInit {
       }
       onDelete(id)
       {
-        this.service.deleteEns(id).subscribe(res=>{
-          this.service.GetAll();
-        }),(err:any)=>{
-          console.log(err);
-        };
+        
+        this.dialogService.openConfirmDialog()
+        .afterClosed().subscribe(res =>{
+          if(res){
+            this.service.deleteEns(id).subscribe(res=>{
+              this.service.GetAll();
+            }),err=>{
+              console.log(err);
+            };
+          }
+        });
       }
-      onDeleteClick(){
-        this.popup.options = {
-          header: "Supprimer ",
-          color: "#5cb85c", // red, blue....
-          widthProsentage: 40, // The with of the popou measured by browser width
-          animationDuration: 1, // in seconds, 0 = no animation
-          showButtons: true, // You can hide this in case you want to use custom buttons
-          confirmBtnContent: "OK", // The text on your confirm button
-          cancleBtnContent: "Cancel", // the text on your cancel button
-          confirmBtnClass: "btn btn-danger", // your class for styling the confirm button
-          cancleBtnClass: "btn btn-info", // you class for styling the cancel button
-          animation: "fadeInDown" // 'fadeInLeft', 'fadeInRight', 'fadeInUp', 'bounceIn','bounceInDown'
-      };
-       
-        this.popup.show(this.popup.options);
+      onOpenEdit(ens){
+        this.service.fillForm(ens);
+        const dialogConfig=new MatDialogConfig();
+        dialogConfig.disableClose=true;
+        dialogConfig.autoFocus=true;
+        dialogConfig.width="50%";
+        dialogConfig.height="95%";
+        this.dialog.open(FormEnseignantComponent,dialogConfig);
+        console.log(ens);
       }
-      onCancel(){
-        this.popup.hide();
-      }
+     
+      
   
 }
